requirements.txt
Flask==3.0.3
flask-cors==4.0.0
flasgger==0.9.7
pymongo==4.6.3
python-dotenv==1.0.1
gunicorn==22.0.0
pytest==8.2.0
mongomock==4.1.2
requests==2.32.3
PyJWT==2.8.0
